Second stage
Moving the code and updating the OS Places address lookup service.
