+++ lib/defra_ruby/address.rb
@@ -4,7 +4,7 @@ require_relative "address/configuration"
 require_relative "address/no_match_error"
 require_relative "address/response"
 
-require_relative "address/services/base_address_service"
+require_relative "address/services/base_service"
 require_relative "address/services/os_places_address_lookup_service"
 
 module DefraRuby
